requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
pydantic
python-multipart